more serializer breakout
