+++ koku/api/report/aws/serializers.py
@@ -15,70 +15,11 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 """Report Serializers."""
-from django.utils.translation import ugettext as _
 from pint.errors import UndefinedUnitError
 from rest_framework import serializers
 
+from api.report.serializers import (StringOrListField, handle_invalid_fields, validate_field)
 from api.utils import UnitConverter
-
-
-def handle_invalid_fields(this, data):
-    """Validate incoming data.
-
-    Args:
-        data    (Dict): data to be validated
-    Returns:
-        (Dict): Validated data
-    Raises:
-        (ValidationError): if field inputs are invalid
-    """
-    unknown_keys = None
-    if hasattr(this, 'initial_data'):
-        unknown_keys = set(this.initial_data.keys()) - set(this.fields.keys())
-    if unknown_keys:
-        error = {}
-        for unknown_key in unknown_keys:
-            error[unknown_key] = _('Unsupported parameter')
-        raise serializers.ValidationError(error)
-    return data
-
-
-def validate_field(this, field, serializer_cls, value):
-    """Validate the provided fields.
-
-    Args:
-        field    (String): the field to be validated
-        serializer_cls (Class): a serializer class for validation
-        value    (Object): the field value
-    Returns:
-        (Dict): Validated value
-    Raises:
-        (ValidationError): if field inputs are invalid
-    """
-    field_param = this.initial_data.get(field)
-    serializer = serializer_cls(data=field_param)
-    serializer.is_valid(raise_exception=True)
-    return value
-
-
-class StringOrListField(serializers.ListField):
-    """Serializer field to handle types that are string or list.
-
-    Converts everything to a list.
-    """
-
-    def to_internal_value(self, data):
-        """Handle string data then call super.
-
-        Args:
-            data    (String or List): data to be converted
-        Returns:
-            (List): Transformed data
-        """
-        list_data = data
-        if isinstance(data, str):
-            list_data = [data]
-        return super().to_internal_value(list_data)
 
 
 class GroupBySerializer(serializers.Serializer):
@@ -291,7 +232,7 @@ class QueryParamSerializer(serializers.Serializer):
         unit_converter = UnitConverter()
         try:
             unit_converter.validate_unit(value)
-        except (AttributeError, UndefinedUnitError) as err:
+        except (AttributeError, UndefinedUnitError):
             error = {'units': f'{value} is not a supported unit'}
             raise serializers.ValidationError(error)
