+++ koku/api/report/serializers.py
@@ -1,0 +1,78 @@
+#
+# Copyright 2018 Red Hat, Inc.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+"""Common serializer logic."""
+from django.utils.translation import ugettext as _
+from rest_framework import serializers
+
+
+def handle_invalid_fields(this, data):
+    """Validate incoming data.
+
+    Args:
+        data    (Dict): data to be validated
+    Returns:
+        (Dict): Validated data
+    Raises:
+        (ValidationError): if field inputs are invalid
+    """
+    unknown_keys = None
+    if hasattr(this, 'initial_data'):
+        unknown_keys = set(this.initial_data.keys()) - set(this.fields.keys())
+    if unknown_keys:
+        error = {}
+        for unknown_key in unknown_keys:
+            error[unknown_key] = _('Unsupported parameter')
+        raise serializers.ValidationError(error)
+    return data
+
+
+def validate_field(this, field, serializer_cls, value):
+    """Validate the provided fields.
+
+    Args:
+        field    (String): the field to be validated
+        serializer_cls (Class): a serializer class for validation
+        value    (Object): the field value
+    Returns:
+        (Dict): Validated value
+    Raises:
+        (ValidationError): if field inputs are invalid
+    """
+    field_param = this.initial_data.get(field)
+    serializer = serializer_cls(data=field_param)
+    serializer.is_valid(raise_exception=True)
+    return value
+
+
+class StringOrListField(serializers.ListField):
+    """Serializer field to handle types that are string or list.
+
+    Converts everything to a list.
+    """
+
+    def to_internal_value(self, data):
+        """Handle string data then call super.
+
+        Args:
+            data    (String or List): data to be converted
+        Returns:
+            (List): Transformed data
+        """
+        list_data = data
+        if isinstance(data, str):
+            list_data = [data]
+        return super().to_internal_value(list_data)
